+++ commands/viewer_commands.py
@@ -11,6 +11,9 @@ from views.rewards.redeem_reward_view import RedeemRewardView
 from threading import Thread
 import requests
 import logging
+import datetime
+import time
+from pytz import timezone
 from config import YAMLConfig as Config
 from discord.app_commands.errors import AppCommandError, CheckFailure
 from util.server_utils import get_base_url
@@ -27,6 +30,11 @@ AUTH_TOKEN = Config.CONFIG["Secrets"]["Server"]["Token"]
 
 # It's stupid that it's here but I don't know how else to make it work
 ACTIVE_CHATTER_KEYWORD = None
+
+PACIFIC_TZ = timezone("US/Pacific")
+# Number representing day of the week, from 0 through to 6
+VOD_REVIEW_DAY = 3
+VOD_REVIEW_DAY_END = 23
 
 
 @app_commands.guild_only()
@@ -104,7 +112,19 @@ class ViewerCommands(app_commands.Group, name="hooj"):
     @app_commands.command(name="submit_vod")
     async def start(self, interaction: Interaction):
         """Opens the VOD Submission Prompt"""
+        today = datetime.datetime.now(PACIFIC_TZ)
+        # Disable submissions on vod review day and inform user what time they can submit again
+        if today.weekday() == VOD_REVIEW_DAY:
+            today = today.replace(hour=VOD_REVIEW_DAY_END, minute=0)
+            vod_review_date = (
+                today + datetime.timedelta((VOD_REVIEW_DAY - today.weekday()) % 7)
+            ).astimezone(timezone("UTC"))
+            unixtime = time.mktime(vod_review_date.timetuple())
 
+            return await interaction.response.send_message(
+                f"No new vods are accepted on vod review day, you can submit again at <t:{unixtime:.0f}:f>",
+                ephemeral=True,
+            )
         modal = NewVodSubmissionModal(self.client)
         await interaction.response.send_modal(modal)